fix(bench): debate arm model_name attribution + de-risk circular import + skip empty-proposal critique
